cnf-tests: Add more features to test

+++ ci-operator/step-registry/telco5g/cnf-tests/telco5g-cnf-tests-commands.sh
@@ -24,6 +24,13 @@ sriov "Test Connectivity Connectivity between client and server Should work over
 EOF
 }
 
+function create_blocking_features_file {
+# List of features that can't run in parallel (affecting the same cluster components)
+cat <<EOF >"${BLOCKING_FEATURES_FILE}"
+sriov dpdk bond vrf tuningcni
+EOF
+}
+
 function create_ns {
     ns=$1
 
@@ -77,9 +84,29 @@ function get_skip_tests {
     echo "${skip_list}"
 }
 
+function get_blocking_features {
+    feature="$1"
+
+    blockers_list=""
+    if [ -f "${BLOCKING_FEATURES_FILE}" ]; then
+        rm -f feature_blockers_list.txt
+        grep --text "${feature}" "${BLOCKING_FEATURES_FILE}" > feature_blockers_list.txt
+        skip_list=""
+        while read line;
+        do
+            blockers=$(echo ${line/${feature}/})
+            blockers_list="${blockers_list} ${blockers}"
+        done < feature_blockers_list.txt
+    fi
+
+    echo "${blockers_list}"
+}
+
 function deploy_and_test {
     feature=$1
     nodes=$2
+
+    touch "${cnf_dir}/testing_${feature}.txt"
 
     # work in tmp dir to be able compile cnf-tests in parallel, per tested feature
     tmp_dir=$(mktemp -d -t cnf-XXXXX)
@@ -97,12 +124,13 @@ function deploy_and_test {
         features_env="typical-baremetal"
     fi
 
-    if [[ "${feature}" == "sriov" ]]; then
-        rm -f perf_profile_for_sriov.yaml
-        oc kustomize feature-configs/typical-baremetal/performance > perf_profile_for_sriov.yaml
-        sed -i "s/name\: performance/name\: performance-sriov/g" perf_profile_for_sriov.yaml
-        sed -i "s/worker-cnf/${node_label}/g" perf_profile_for_sriov.yaml
-        oc apply -f perf_profile_for_sriov.yaml
+    if [[ "${feature}" == "sriov" ]] || [[ "${feature}" == "dpdk" ]]; then
+        perf_profile_file="perf_profile_for_${feature}.yaml"
+        rm -f "${perf_profile_file}"
+        oc kustomize feature-configs/typical-baremetal/performance > "${perf_profile_file}"
+        sed -i "s/name\: performance/name\: performance-${feature}/g" "${perf_profile_file}"
+        sed -i "s/worker-cnf/${node_label}/g" "${perf_profile_file}"
+        oc apply -f "${perf_profile_file}"
     fi
 
     if [[ "${feature}" == "xt_u32" ]] || [[ "${feature}" == "sctp" ]]; then
@@ -116,6 +144,11 @@ function deploy_and_test {
 
     CNF_NODES="${nodes}" make setup-test-cluster
 
+    if [[ "${feature}" == "dpdk" ]]; then
+        FEATURES_ENVIRONMENT="${features_env}" FEATURES="sriov" make feature-deploy
+        FEATURES_ENVIRONMENT="${features_env}" FEATURES="sriov general" make feature-wait
+    fi
+
     FEATURES_ENVIRONMENT="${features_env}" FEATURES="${feature}" make feature-deploy
     FEATURES_ENVIRONMENT="${features_env}" FEATURES="${feature} general" make feature-wait
 
@@ -127,10 +160,13 @@ function deploy_and_test {
         oc label "${node}" "node-role.kubernetes.io/${node_label}"-
         touch "${cnf_dir}/${node}_ready.txt"
     done
+
+    rm -f "${cnf_dir}/testing_${feature}.txt"
 }
 
-export FEATURES="${FEATURES:-sriov performance sctp xt_u32 ovn}" # next: ovs_qos metallb
+export FEATURES="${FEATURES:-sriov dpdk performance sctp xt_u32 ovn}" # next: ovs_qos metallb
 export SKIP_TESTS_FILE="${SKIP_TESTS_FILE:-${SHARED_DIR}/telco5g-cnf-tests-skip-list.txt}"
+export BLOCKING_FEATURES_FILE="${BLOCKING_FEATURES_FILE:-${SHARED_DIR}/telco5g-cnf-tests-blocking-features-list.txt}"
 export SCTPTEST_HAS_NON_CNF_WORKERS="${SCTPTEST_HAS_NON_CNF_WORKERS:-false}"
 export XT_U32TEST_HAS_NON_CNF_WORKERS="${XT_U32TEST_HAS_NON_CNF_WORKERS:-false}"
 
@@ -174,32 +210,55 @@ make setup-build-index-image
 cd -
 
 create_tests_skip_list_file
+create_blocking_features_file
 
 # run cnf-tests by feature in a thread on a free worker node
-for feature in ${FEATURES}; do
-    log_file="${ARTIFACT_DIR}/deploy_and_test_${feature}.log"
-    rm -f "${log_file}"
-    
-    feature_nodes=""
-    num_of_free_nodes=0
-    num_of_required_nodes=1
-    if [[ "${feature}" == "ovs_qos" ]]; then
-        num_of_required_nodes=2
-    fi
-    while [ ${num_of_free_nodes} -lt ${num_of_required_nodes} ]; do
-        for node in ${test_nodes}; do
-            node_ready_file="${node}_ready.txt"
-            if [ -f "${node_ready_file}" ]; then
-                rm -f "${node_ready_file}"
-                feature_nodes="${feature_nodes} ${node}"
-                num_of_free_nodes=$((num_of_free_nodes+1))
-            fi
-            if [ ${num_of_free_nodes} -eq ${num_of_required_nodes} ]; then
-                (deploy_and_test "${feature}" "${feature_nodes}" || true) 2>&1 | tee "${log_file}" &
-                break
+num_of_features=$(echo "${FEATURES}" | wc -w)
+num_of_tested_features=0
+while [ "${num_of_tested_features}" -lt "${num_of_features}" ]; do
+    for feature in ${FEATURES}; do
+        if [ -f "${cnf_dir}/${feature}_tested.txt" ]; then
+            continue
+        fi
+
+        blocking_features=$(get_blocking_features "${feature}")
+        blocked=false
+        for blocker in ${blocking_features}; do
+            if [ -f "${cnf_dir}/testing_${blocker}.txt" ]; then
+                blocked=true
             fi
         done
-        sleep 10
+        if ${blocked}; then
+            continue
+        fi
+            
+
+        log_file="${ARTIFACT_DIR}/deploy_and_test_${feature}.log"
+        rm -f "${log_file}"
+        
+        feature_nodes=""
+        num_of_free_nodes=0
+        num_of_required_nodes=1
+        if [[ "${feature}" == "ovs_qos" ]]; then
+            num_of_required_nodes=2
+        fi
+        while [ ${num_of_free_nodes} -lt ${num_of_required_nodes} ]; do
+            for node in ${test_nodes}; do
+                node_ready_file="${node}_ready.txt"
+                if [ -f "${node_ready_file}" ]; then
+                    rm -f "${node_ready_file}"
+                    feature_nodes="${feature_nodes} ${node}"
+                    num_of_free_nodes=$((num_of_free_nodes+1))
+                fi
+                if [ ${num_of_free_nodes} -eq ${num_of_required_nodes} ]; then
+                    (deploy_and_test "${feature}" "${feature_nodes}" || true) 2>&1 | tee "${log_file}" &
+                    touch "${cnf_dir}/${feature}_tested.txt"
+                    num_of_tested_features=$((num_of_tested_features+1))
+                    break
+                fi
+            done
+            sleep 10
+        done
     done
 done
 wait
